refactor: some changes
